fix: Bug card faq to faq page
