+++ src/components/index.ts
@@ -5,4 +5,4 @@ export * from './ui/top-menu/TopMenu'
 export * from './ui/mamuts/badges/BadgeDate'
 export * from './ui/mamuts/buttons/ButtonSeMore'
 export * from './ui/mamuts/badges/BadgeCategory'
-export * from './ui/fap/CardFaq'
+export * from './ui/faq/CardFaq'
