+++ src/data/data-faq/data-faq.ts
@@ -1,20 +1,20 @@
 export const DataFaq = [
-	{
-		id: 1,
-		question: 'Si soy  mujer, ¿Puedo ir a los compas de Hierro?',
-		response:
-			'Claro que sí, en los compas de Hierro no hay distinción de género, todos son bienvenidos.'
-	},
-	{
-		id: 2,
-		question: '¿Cuanto cuesta ir a los compas de hierro?',
-		response:
-			'Nada, los compas de hierro es una organizacion si fines de lucro solo buscamos gente comprometida en el modo guerra'
-	},
-	{
-		id: 3,
-		question: '¿Que debo hacer si mi novia me pide un tiempo?',
-		response:
-			'"Tiempo" palabra  usada por una mujer para comunicar de manera indirecta que quiere irse con alguien mas pero no quiere que le dejes de hablar, en conclusion salga de ahi mi compa.'
-	}
+  {
+    id: 1,
+    question: 'Si soy  mujer, ¿Puedo ir a los compas de Hierro?',
+    response:
+      'Claro que sí, en los compas de Hierro no hay distinción de género, todos son bienvenidos.',
+  },
+  {
+    id: 2,
+    question: '¿Cuanto cuesta ir a los compas de hierro?',
+    response:
+      'Nada, los compas de hierro es una organizacion si fines de lucro solo buscamos gente comprometida en el modo guerra',
+  },
+  {
+    id: 3,
+    question: '¿Que debo hacer si mi novia me pide un tiempo?',
+    response:
+      '"Tiempo" palabra  usada por una mujer para comunicar de manera indirecta que quiere irse con alguien mas pero no quiere que le dejes de hablar, en conclusion salga de ahi mi compa.',
+  },
 ]
